api endpoints hardened, renamed files
hardened enpoints where id's could be manipulated and potentially delete posts that are not their own. allow owner to comment on posts that are not allowed to be commented on by the post settings themselves unless they own it. Added logic to prevent users from commenting on a posts setting that specifically have the commenting or other settins disabled.