updated backend
fixed logical errors, made api clearer to understand with the "if()" checks, added "v_" to all variables for the validation middleware to prevent variable conflicts, removed dayjs in favor of the date.now(). set the hgetall reply to be null if the key isn't right instead of an empty array.